move all encoding methods to encoding.go
