+++ types/block.go
@@ -3,14 +3,7 @@ package types
 // block.go defines the Block type for Sia, and provides some helper functions
 // for working with blocks.
 
-import (
-	"encoding/json"
-	"fmt"
-	"io"
-
-	"github.com/NebulousLabs/Sia/crypto"
-	"github.com/NebulousLabs/Sia/encoding"
-)
+import "github.com/NebulousLabs/Sia/crypto"
 
 const (
 	// BlockHeaderSize is the size, in bytes, of a block header.
@@ -131,37 +124,4 @@ func (b Block) MinerPayoutID(i uint64) SiacoinOutputID {
 		b.ID(),
 		i,
 	))
-}
-
-// MarshalSia implements the encoding.SiaMarshaler interface.
-func (b Block) MarshalSia(w io.Writer) error {
-	w.Write(b.ParentID[:])
-	w.Write(b.Nonce[:])
-	encoding.WriteUint64(w, uint64(b.Timestamp))
-	return encoding.NewEncoder(w).EncodeAll(b.MinerPayouts, b.Transactions)
-}
-
-// UnmarshalSia implements the encoding.SiaUnmarshaler interface.
-func (b *Block) UnmarshalSia(r io.Reader) error {
-	io.ReadFull(r, b.ParentID[:])
-	io.ReadFull(r, b.Nonce[:])
-	tsBytes := make([]byte, 8)
-	io.ReadFull(r, tsBytes)
-	b.Timestamp = Timestamp(encoding.DecUint64(tsBytes))
-	return encoding.NewDecoder(r).DecodeAll(&b.MinerPayouts, &b.Transactions)
-}
-
-// MarshalJSON marshales a block id as a hex string.
-func (bid BlockID) MarshalJSON() ([]byte, error) {
-	return json.Marshal(bid.String())
-}
-
-// String prints the block id in hex.
-func (bid BlockID) String() string {
-	return fmt.Sprintf("%x", bid[:])
-}
-
-// UnmarshalJSON decodes the json hex string of the block id.
-func (bid *BlockID) UnmarshalJSON(b []byte) error {
-	return (*crypto.Hash)(bid).UnmarshalJSON(b)
 }
