+++ types/transactions_test.go
@@ -1,7 +1,6 @@
 package types
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/NebulousLabs/Sia/crypto"
@@ -66,30 +65,5 @@ func TestTransactionSiacoinOutputSum(t *testing.T) {
 	}
 	if txn.SiacoinOutputSum().Cmp(NewCurrency64(654321)) != 0 {
 		t.Error("wrong siacoin output sum was calculated, got:", txn.SiacoinOutputSum())
-	}
-}
-
-// TestSpecifierMarshaling tests the marshaling methods of the specifier
-// type.
-func TestSpecifierMarshaling(t *testing.T) {
-	s1 := SpecifierClaimOutput
-	b, err := json.Marshal(s1)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var s2 Specifier
-	err = json.Unmarshal(b, &s2)
-	if err != nil {
-		t.Fatal(err)
-	} else if s2 != s1 {
-		t.Fatal("mismatch:", s1, s2)
-	}
-
-	// invalid json
-	x := 3
-	b, _ = json.Marshal(x)
-	err = json.Unmarshal(b, &s2)
-	if err == nil {
-		t.Fatal("Unmarshal should have failed")
 	}
 }
